Policy Contacts

+++ app/Http/Controllers/ContactController.php
@@ -5,7 +5,8 @@ namespace App\Http\Controllers;
 use App\Models\Contact;
 
 use Illuminate\Http\Request;
-
+use Illuminate\Http\Response;
+use Illuminate\Support\Facades\Gate;
 
 class ContactController extends Controller
 {
@@ -16,7 +17,8 @@ class ContactController extends Controller
      */
     public function index()
     {
-        return view('contacts.index', ['contacts' => Contact::all()]);
+        $contacts = auth()->user()->contacts;
+        return view('contacts.index', compact('contacts'));
     }
 
     /**
@@ -44,8 +46,7 @@ class ContactController extends Controller
             'age' => 'required | numeric | min:1 | max:255',
         ]);
         
-        Contact::create($data);
-         
+       auth()->user()->contacts()->create($data);
         return redirect()->route('home');
     }
 
@@ -57,6 +58,7 @@ class ContactController extends Controller
      */
     public function show(Contact $contact)
     {
+        $this->authorize('view', $contact);
         return view('contacts.show',compact('contact'));
     }
 
@@ -68,6 +70,7 @@ class ContactController extends Controller
      */
     public function edit(Contact $contact)
     {
+        $this->authorize('update', $contact);
         return view ('contacts.edit', compact('contact'));
     }
 
@@ -80,6 +83,8 @@ class ContactController extends Controller
      */
     public function update(Request $request, Contact $contact)
     {
+        $this->authorize('update', $contact);
+
         $data = $request->validate([
             'name' => 'required',
             'email' => 'required | email',
@@ -100,7 +105,10 @@ class ContactController extends Controller
      */
     public function destroy(Contact $contact)
     {
-        $contact->delete(); 
+        $this->authorize('delete', $contact);
+
+        $contact->delete();
+
         return redirect()->route('home');   
     }
 }
